# openapi.yaml
openapi: 3.0.1
info:
  title: Smart Agenda API
  description: API para una aplicación de gestión de tareas inteligente.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
...
paths:
  /users/register:
    post:
      tags:
        - Users
      summary: Registro de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
                - correo
                - contraseña
              properties:
                nombre:
                  type: string
                correo:
                  type: string
                contraseña:
                  type: string
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
        '400':
          description: Error al crear usuario
  /users/login:
    post:
      tags:
        - Users
      summary: Inicio de sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - correo
                - contraseña
              properties:
                correo:
                  type: string
                contraseña:
                  type: string
      responses:
        '200':
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Credenciales inválidas
        '404':
          description: Usuario no encontrado
  /users/profile:
    get:
      tags:
        - Users
      summary: Obtener perfil de usuario
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil de usuario obtenido exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  nombre:
                    type: string
                  correo:
                    type: string
        '401':
          description: No autorizado

    put:
      tags:
        - Users
      summary: Actualizar perfil de usuario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                correo:
                  type: string
                contraseña:
                  type: string
      responses:
        '200':
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile updated successfully
        '401':
          description: No autorizado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT